Skip products not in stock
Not all products are in stock and those that are not are not included in
the inital reponse from the server. The server expects to get a list
back of all products but this is not the case anymore. To resolve this
we only look at products that are in the reponse from the client.